+++ tests/csv_test.py
@@ -11,3 +11,6 @@ from yabc import csv_to_json
 class CsvTest(unittest.TestCase):
     def test_load_coinbase_csv(self):
         csv_to_json.coinbase_to_dict("testdata/synthetic_coinbase_csv.csv")
+
+    def test_load_gemini_csv(self):
+        csv_to_json.gemini_to_dict("testdata/synthetic_gemini_csv.csv")
